third_party/blink/web_tests/external/wpt/web-nfc/NDEFWriter_push.https-expected.txt

This is a testharness.js-based test.
PASS Test that promise is rejected with TypeError if NDEFMessageSource is invalid.
PASS Test that promise is rejected with SyntaxError if NDEFMessageSource contains invalid records.
PASS NDEFWriter.push should fail if abort push request before push happends.
PASS NDEFWriter.push should fail if signal's aborted flag is set.
PASS NDEFWriter.push should fail if signal is not an AbortSignal.
PASS Synchronously signaled abort.
PASS NDEFWriter.push should fail when NFC HW is disabled.
PASS NDEFWriter.push should fail when NFC HW is not supported.
PASS Reject promise with NotSupportedError if NFC message size exceeds 32KB.
PASS Reject promise with SyntaxError if WebNFC Id cannot be created from provided URL.
PASS NDEFWriter.push should fail with TypeError when invalid target value is provided.
PASS Test that WebNFC API is not accessible from iframe context.
PASS NDEFWriter.push should succeed when NFC HW is enabled
PASS NDEFWriter.push NDEFMessage containing text, mime, unknown, url, absolute-url and external records with default NDEFPushOptions.
PASS Test that NDEFWriter.push succeeds when message is DOMString.
PASS Test that NDEFWriter.push succeeds when message is ArrayBuffer.
PASS Test that NDEFWriter.push succeeds when message is ArrayBufferView.
PASS NDEFWriter.push with 'empty' record should succeed.
PASS Check that default NDEFPushOptions values are correctly set.
PASS Check that provided NDEFPushOptions values are correctly converted.
PASS NDEFWriter.push should read data when ignoreRead is false.
PASS NDEFWriter.push should ignore reading data when ignoreRead is true.
PASS NDEFWriter.push should replace all previously configured push operations.
PASS Test that recordType should be set to 'text' if NDEFRecordInit.record's recordType is undefined and NDEFRecordInit.record's data is DOMString.
PASS Test that recordType should be set to 'mime' if NDEFRecordInit.record's recordType is undefined and NDEFRecordInit.record's data is not DOMString.
PASS Test that mediaType should be set to 'application/octet-stream' if NDEFRecordInit.record's recordType is 'mime' and NDEFRecordInit.record's mediaType is undefined.
PASS NDEFWriter.push should fail when the NFC device does not expose NDEF technology.
PASS NDEFWriter.push should succeed to push data to an unformatted NFC device when the NDEFPushOptions.overwrite is false.
PASS NDEFWriter.push should succeed to overwrite the existing data when the NDEFPushOptions.overwrite is true.
FAIL NDEFWriter.push should fail when there are NDEF records on the NFC device and NDEFPushOptions.overwrite is false. assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
